Renamed `callback` endpoint to `spotify-auth`
## What
- Renamed `callback` endpoint to `spotify-auth`

## Why
- Align with new spotify callback endpoint name, renamed for better description of function purpose

## Impact
- Will customer experience be significantly altered? No
- Does the code change cover more than one task? No
